+++ backend/contracts/CampaignFactory.sol
@@ -15,11 +15,11 @@ contract CampaignFactory is Ownable {
         _;
     }
 
-    function createCrowdfundingCampaign(string memory _uri, address _artistAddr, string memory _name, uint8 _fees, string memory _description, uint8 _nbTiers) external isContractOwner returns (address campaignAddress) {
+    function createCrowdfundingCampaign(string memory _uri, address _artistAddr, string memory _name, uint8 _fees, string memory _description, uint8 _nbTiers, address _usdcAddr) external isContractOwner returns (address campaignAddress) {
         bytes32 salt = keccak256(abi.encodePacked(_uri, _artistAddr));
         bytes memory campaignBytecode = abi.encodePacked(
             type(CrowdfundingCampaign).creationCode,
-            abi.encode(_uri, _ownerContractAddr, _artistAddr, _name, _fees, _description, _nbTiers)
+            abi.encode(_uri, _ownerContractAddr, _usdcAddr, _artistAddr, _name, _fees, _description, _nbTiers)
         );
 
         assembly {