+++ backend/contracts/CrowdfundingCampaign.sol
@@ -5,6 +5,7 @@ import '../node_modules/@openzeppelin/contracts/token/ERC1155/ERC1155.sol';
 import '../node_modules/@openzeppelin/contracts/token/ERC1155/extensions/ERC1155Burnable.sol';
 import '../node_modules/@openzeppelin/contracts/token/ERC1155/extensions/ERC1155Supply.sol';
 import '../node_modules/@openzeppelin/contracts/utils/Strings.sol';
+import '../node_modules/@openzeppelin/contracts/token/ERC20/IERC20.sol';
 
 contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
     string  private _baseUri;
@@ -13,6 +14,7 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
     bool    private _campaignInProgress;
     uint8   private _nbFilledTiers;
     address private _tuneTogetherAddr;
+    IERC20  private _usdc;
 
     string  public name;
     string  public description;
@@ -20,12 +22,13 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
     address public artistAddress;
     uint8   public nbTiers;
 
-    mapping(uint8 => uint) tierPrices;
+    mapping(uint8 => uint) private tierPrices;
 
     event CampaignStarted(uint256 _startTimestamp, uint256 _endTimestamp);
     event CampaignClosed(uint256 _endTimestamp);
     event CampaignInfoUpdated(string _name, string _description, uint8 _fees);
     event TierPriceAdded(uint8 _id, uint _price);
+    event FundWithdraw(address _artistAddr, uint _usdcBalance, uint _ethBalance, uint _timestamp);
 
     modifier isContractOwner() {
         require(_tuneTogetherAddr == msg.sender, 'You\'re not the owner');
@@ -49,9 +52,16 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
         _;
     }
 
-    constructor(string memory baseUri_, address tuneTogetherAddr_, address _artistAddr, string memory _name, uint8 _fees, string memory _description, uint8 _nbTiers) ERC1155(baseUri_) {
+    modifier campaignClosed() {
+        require(_campaignStarted, 'Artist didn\'t start the campaign yet');
+        require(!_campaignInProgress || _endTimestamp < block.timestamp, 'Campaign in progress');
+        _;
+    }
+
+    constructor(string memory baseUri_, address tuneTogetherAddr_, address usdcAddr_, address _artistAddr, string memory _name, uint8 _fees, string memory _description, uint8 _nbTiers) ERC1155(baseUri_) {
         _baseUri = baseUri_;
         _tuneTogetherAddr = tuneTogetherAddr_;
+        _usdc = IERC20(usdcAddr_);
         artistAddress = _artistAddr;
         name = _name;
         fees = _fees;
@@ -59,13 +69,15 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
         nbTiers = _nbTiers;
     }
 
-    function mint(uint8 _id, uint256 _amount) public campaignInProgress payable {
-        require(msg.value == _amount * tierPrices[_id], 'Wrong value');
+    function mint(uint8 _id, uint256 _amount) public campaignInProgress {
+        require(_usdc.balanceOf(msg.sender) >= _amount * tierPrices[_id], "Not enough balance");
+
+        _usdc.transferFrom(msg.sender, address(this), _amount * tierPrices[_id]);
         _mint(msg.sender, _id, _amount, '');
     }
 
     function uri(uint _tokenId) public view override returns (string memory) {
-        return string.concat(_baseUri, Strings.toString(_tokenId), '.json');
+        return string(abi.encodePacked(_baseUri, Strings.toString(_tokenId), '.json'));
     }
 
     function startCampaign() external onlyArtist campaignNotStarted {
@@ -100,7 +112,7 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
 
     function setTierPrice(uint8 _id, uint _price) external onlyArtist campaignNotStarted {
         require(_id > 0 && _id <= nbTiers, 'Tier does not exist');
-        require(_price >= 0.00001 ether, 'Price too low');
+        require(_price >= 1 * 10**6 , 'Price too low');
 
         if (_id > 1) {
             require(_price > tierPrices[_id - 1], 'Price should be higher than the previous tier');
@@ -114,6 +126,24 @@ contract CrowdfundingCampaign is ERC1155, ERC1155Burnable, ERC1155Supply {
         _nbFilledTiers = _nbFilledTiers + 1;
 
         emit TierPriceAdded(_id, _price);
+    }
+
+    function getTierPrice(uint8 _id) external view returns (uint _price) {
+        return tierPrices[_id];
+    }
+
+    function withdraw() external onlyArtist campaignClosed {
+        uint ethBalance = address(this).balance;
+        uint usdcBalance = _usdc.balanceOf(address(this));
+
+        _usdc.transfer(msg.sender, usdcBalance);
+        
+        if (ethBalance > 0) {
+            (bool success, ) = msg.sender.call{value: ethBalance}('');
+            require(success, "Withdraw failed");
+        }
+
+        emit FundWithdraw(msg.sender, usdcBalance, ethBalance, block.timestamp);
     }
 
     // The following functions are overrides required by Solidity.